Update parameter callback

+++ pcl_detector/include/pcl_detector/detectors/dbscan_detector.hpp
@@ -30,7 +30,7 @@ public:
 
 private:
     float m_eps;
-    int m_min_points;
+    size_t m_min_points;
 };
 
 }; // namespace pcl_detector
